refactor(redux): first state variables managed with redux
- moved some variables to redux
- created ModSelector component
