+++ package.json
@@ -3,16 +3,21 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@reduxjs/toolkit": "^1.2.5",
+    "axios": "^0.21.1",
     "bootstrap": "^4.3.1",
     "jquery": "^3.4.1",
-    "popper.js": "^1.15.0",
-    "react": "^16.4.1",
-    "react-bootstrap": "^0.32.1",
-    "react-dom": "^16.4.1",
-    "react-form": "^3.5.6",
-    "react-scripts": "^3.1.1",
+    "lodash": "^4.17.21",
     "pako": "^2.0.3",
-    "axios": "^0.21.1"
+    "popper.js": "^1.15.0",
+    "react": "^17.0.1",
+    "react-bootstrap": "^0.32.1",
+    "react-dom": "^17.0.1",
+    "react-form": "^3.5.6",
+    "react-redux": "^7.2.2",
+    "react-scripts": "^3.1.1",
+    "redux": "^4.0.5",
+    "redux-thunk": "^2.3.0"
   },
   "scripts": {
     "start": "PORT=3001 BROWSER=none REACT_APP_URLROOT=https://reports.alliancegenome.org/ react-scripts start",
