+++ src/redux/selectors.js
@@ -1,0 +1,3 @@
+export const getSelectedMod = store => store ? store.selectedMod : undefined;
+export const getModsList = store => store ? store.modsList : [];
+export const getLatestFilesOnly = store => store ? store.latestFilesOnly : true;
