+++ src/redux/store.js
@@ -1,0 +1,6 @@
+import { createStore, applyMiddleware } from "redux";
+import thunk from 'redux-thunk';
+import reportToolReducer from "./reducers";
+
+const store = createStore(reportToolReducer, applyMiddleware(thunk));
+export default store;
